Fix incorrect status on domain failure log
This was hardcoded to be 500, which was incorrect. It is erroring becuase it is not 200. Simplified it by only passing status code and not pre-creating any kind of error message, we create the message with the status code when we need to ourselves, in this case in the send email functions